:hammer: Refactor code: Move google login from app.py to login package
