--- WEB-CONVERTER/src/com/jalasoft/Web-converter/blueprints/login_bp.py
@@ -1,21 +1,0 @@
-# Python standard libraries
-import json
-import os
-import sqlite3
-from dotenv import load_dotenv
-load_dotenv()
-
-# Third party libraries
-from flask import Flask, redirect, request, url_for
-from flask_login import (
-    LoginManager,
-    current_user,
-    login_required,
-    login_user,
-    logout_user,
-)
-from oauthlib.oauth2 import WebApplicationClient
-import requests
-
-
-
