+++ WEB-CONVERTER/src/com/jalasoft/Web-converter/blueprints/video_to_video_bp.py
@@ -15,7 +15,7 @@ from flask import Flask
 from flask import render_template
 from blueprints.handle_inputs import HandleInputs
 from blueprints.converter_base_bp import ConverterBase
-from module.user_authenticate import LoggedUser
+from login.user_authenticate import LoggedUser
 
 # app = Flask(__name__)
 video_to_video_blueprint = Blueprint('video_to_video', __name__)
