+++ WEB-CONVERTER/src/com/jalasoft/Web-converter/templates/base.html
@@ -77,6 +77,9 @@
         </header>
     </div>
     <div class="container py-3"> {% block body %} {% endblock %} </div>
+    <footer>
+      <p>Copyright © 2023 Jalasoft - All rights reserved</p>
+    </footer>
     <script src="static/js/styles.js"></script>
     <script src="static/js/animations.js"></script>
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
@@ -85,7 +88,4 @@
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"> </script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.min.js" integrity="sha384-w1Q4orYjBQndcko6MimVbzY0tgp4pWB4lZ7lr30WKz0vr/aWKhXdBNmNb5D92v7s" crossorigin="anonymous"> </script>
 </body>
-<footer>
-  <p>Copyright © 2023 Jalasoft - All rights reserved</p>
-</footer>
 </html>
