+++ WEB-CONVERTER/src/com/jalasoft/Web-converter/userDB/user.py
@@ -1,6 +1,6 @@
 from flask_login import UserMixin
 
-from db import get_db
+from userDB.db import get_db
 
 
 class User(UserMixin):
